Reorganize issue modules
