+++ tests/dataset.spec.js
@@ -1,7 +1,7 @@
 const assert = require('chai').assert
 const hed = require('../validator/dataset')
 const schema = require('../validator/schema')
-const generateValidationIssue = require('../utils/issues').generateIssue
+const generateValidationIssue = require('../utils/issues/issues').generateIssue
 const generateConverterIssue = require('../converter/issues')
 
 describe('HED dataset validation', () => {
